Switch from using clean-css internally to executing it as a separate program
This is so that we can set another program off and wait for it in an async
manor, instead of doing the computation within the main control loop. Mainly
because, if we hit a CSS file which can cause an infinite loop in clean-css
we'll break the program and won't respond to anything, whereas if it's in a
child process, we can tell execFile() to timeout after a number of seconds,
which means we're not waiting forever or tieing up the main event loop.

Finally! Yes, this has been a long time coming.